+++ src/config/config.ts
@@ -25,6 +25,7 @@ import {sep} from 'path'
 const debug = Debug()
 
 const _pjson = requireJson<PJSON>(__dirname, '..', '..', 'package.json')
+const BASE = `${_pjson.name}@${_pjson.version}`
 
 function channelFromVersion(version: string) {
   const m = version.match(/[^-]+(?:-([^.]+))?/)
@@ -69,7 +70,7 @@ class Permutations extends Map<string, Set<string>> {
 }
 
 export class Config implements IConfig {
-  private _base = `${_pjson.name}@${_pjson.version}`
+  private _base = BASE
 
   public arch!: ArchTypes
   public bin!: string
@@ -111,7 +112,23 @@ export class Config implements IConfig {
 
   private _commandIDs!: string[]
 
-  constructor(public options: Options) {}
+  constructor(public options: Options) {
+    if (options.config) {
+      if (Array.isArray(options.config.plugins) && Array.isArray(this.plugins)) {
+        // incoming config is v2 with plugins array and this config is v2 with plugins array
+        Object.assign(this, options.config)
+      } else if (Array.isArray(options.config.plugins) && !Array.isArray(this.plugins)) {
+        // incoming config is v2 with plugins array and this config is v3 with plugin Map
+        Object.assign(this, options.config, {plugins: new Map(options.config.plugins.map(p => [p.name, p]))})
+      } else if (!Array.isArray(options.config.plugins) && Array.isArray(this.plugins)) {
+        // incoming config is v3 with plugin Map and this config is v2 with plugins array
+        Object.assign(this, options.config, {plugins: options.config.getPluginsList()})
+      } else {
+        // incoming config is v3 with plugin Map and this config is v3 with plugin Map
+        Object.assign(this, options.config)
+      }
+    }
+  }
 
   static async load(opts: LoadOptions = module.filename || __dirname): Promise<Config> {
     // Handle the case when a file URL string is passed in such as 'import.meta.url'; covert to file path.
@@ -120,7 +137,28 @@ export class Config implements IConfig {
     }
 
     if (typeof opts === 'string') opts = {root: opts}
-    if (isConfig(opts)) return opts
+    if (isConfig(opts)) {
+      const {lt} = await import('semver')
+
+      const currentConfigBase = BASE.replace('@oclif/core@', '')
+      const incomingConfigBase = opts._base.replace('@oclif/core@', '')
+      /**
+       * Reload the Config based on the version required by the command.
+       * This is needed because the command is given the Config instantiated
+       * by the root plugin, which may be a different version than the one
+       * required by the command.
+       *
+       * Doing this ensures that the command can freely use any method on Config that
+       * exists in the version of Config required by the command but may not exist on the
+       * root's instance of Config.
+       */
+      if (lt(incomingConfigBase, currentConfigBase)) {
+        debug(`reloading config from ${opts._base} to ${BASE}`)
+        return new Config({...opts.options, config: opts})
+      }
+
+      return opts
+    }
 
     const config = new Config(opts)
     await config.load()
@@ -129,6 +167,8 @@ export class Config implements IConfig {
 
   // eslint-disable-next-line complexity
   public async load(): Promise<void> {
+    if (this.options.config) return
+
     settings.performanceEnabled = (settings.performanceEnabled === undefined ? this.options.enablePerf : settings.performanceEnabled) ?? false
     const plugin = new Plugin.Plugin({root: this.options.root})
     await plugin.load()
@@ -542,6 +582,10 @@ export class Config implements IConfig {
     const url = new URL(host)
     url.pathname = path.join(url.pathname, key)
     return url.toString()
+  }
+
+  public getPluginsList(): IPlugin[] {
+    return this.plugins
   }
 
   protected dir(category: 'cache' | 'data' | 'config'): string {